Mediapipe_GenAI/requirements.txt
# Core
opencv-python
mediapipe
python-dotenv
tqdm
Pillow

# AI / ML Frameworks
# Note: For CUDA support on Windows, please install torch manually:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu124
torch
torchvision

# LangChain & Gemini
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-google-community
langchain-google-genai
google-auth
google-auth-oauthlib
google-auth-httplib2